helpers: added Config helper and moved some functions from MiscFunctions
