+++ src/debug/HyprCtl.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <fstream>
-#include "../helpers/MiscFunctions.hpp"
 #include "../helpers/defer/Promise.hpp"
 #include "../desktop/view/Window.hpp"
 #include "../desktop/DesktopTypes.hpp"
