+++ src/managers/KeybindManager.cpp
@@ -15,6 +15,7 @@
 #include "../managers/input/InputManager.hpp"
 #include "../layout/LayoutManager.hpp"
 #include "../event/EventBus.hpp"
+#include "../helpers/MiscFunctions.hpp"
 
 #include <string>
 #include <cstring>
