+++ src/managers/input/trackpad/gestures/SpecialWorkspaceGesture.cpp
@@ -3,6 +3,7 @@
 #include "../../../../Compositor.hpp"
 #include "../../../../desktop/state/FocusState.hpp"
 #include "../../../../render/Renderer.hpp"
+#include "../../../../helpers/Config.hpp"
 
 #include <hyprutils/memory/Casts.hpp>
 using namespace Hyprutils::Memory;
