+++ src/plugins/HookSystem.cpp
@@ -2,7 +2,6 @@
 #include "../debug/log/Logger.hpp"
 #include "../helpers/varlist/VarList.hpp"
 #include "../managers/TokenManager.hpp"
-#include "../helpers/MiscFunctions.hpp"
 
 #define register
 #include <udis86.h>
